Split tests of rule enforce-store-naming-convention to prefix and postfix cases
