--- rules/enforce-store-naming-convention/examples/correct-store-naming.js
@@ -1,5 +1,0 @@
-import { createStore } from "effector";
-
-const $sum = createStore(0);
-
-export { $sum };
