--- rules/enforce-store-naming-convention/examples/incorrect-store-naming-in-domain.js
@@ -1,7 +1,0 @@
-import { createDomain } from "effector";
-
-const d = createDomain();
-
-const sum = d.createStore(0);
-
-export { sum };
