+++ rules/enforce-store-naming-convention/postfix/examples/incorrect-store-naming-in-domain-with-handlers.js
