+++ rules/enforce-store-naming-convention/postfix/examples/incorrect-store-naming-with-handlers.js
