+++ rules/enforce-store-naming-convention/prefix/examples/correct-store-naming-in-domain-with-handlers.js
