+++ rules/enforce-store-naming-convention/prefix/examples/correct-store-naming-with-handlers.js
